scripts/requirements.txt
# Buffett Pick Data Collection Script Dependencies
# Install: pip install -r requirements.txt

# HTTP requests
requests>=2.31.0

# Data processing
pandas>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Supabase client
supabase>=2.0.0

# Progress bar
tqdm>=4.66.0

# HTML parsing (pandas.read_html dependency)
lxml>=4.9.0
html5lib>=1.1

# yfinance data source
yfinance>=0.2.30
curl_cffi>=0.6.0

# Crypto exchange API (Binance OHLC backfill)
ccxt>=4.0.0
